분리된 컨트롤러를 하나로 합쳐라
종명님이 만드신 `com.kurly.promotion.controller`와
`com.kurly.promotion.ui`가 두개로 나뉘어져서 `com.kurly.promotion.ui`로
합쳤습니다.
